+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to MiniJinja are documented here.
 ## Unreleased
 
 * Added `Environment.SetUnknownMethodCallback` to the Go port.  #934
+* Changed floor division and modulo in Rust and Go to match Jinja2 semantics.  #935
 * Fixed empty iterators in the Go port being incorrectly treated as non-iterable and truthy.  #933
 
 ## 2.24.0